requirements.txt
# Core
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Data Sources (FREE)
openbb>=4.0.0
yfinance>=0.2.0
fredapi>=0.5.0
requests>=2.31.0

# Visualization
plotly>=5.18.0
streamlit>=1.28.0

# AI
ollama>=0.1.0

# Database
sqlalchemy>=2.0.0
chromadb>=0.4.0

# Math/Finance
pysabr>=0.2.0
statsmodels>=0.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Dev
black>=23.0.0
isort>=5.12.0
